+++ init.lua
@@ -4,9 +4,7 @@ local realm = GetRealmName();
 local char = UnitName('player');
 
 function initPulse ()
-	if (Pulse == nil) then
-		Pulse = {};
-	elseif (Pulse.version == 1) then
+	if ((Pulse == nil) or (Pulse.version < 3)) then
 		Pulse = {};
 	end
 	if (Pulse.realm == nil) then
@@ -22,19 +20,25 @@ function initPulse ()
 		Pulse.realm[realm].char[char] = {};
 	end
 
-	Pulse.version = 2;
+	Pulse.version = 3;
 end
 
 frame:RegisterEvent('PLAYER_LOGIN');
 frame:SetScript('OnEvent', function (self, event, ...)
 	if (event == 'PLAYER_LOGIN') then
 
+		initPulse();
+
 		frame:RegisterEvent('BAG_UPDATE');
 		frame:RegisterEvent('ITEM_UNLOCKED');
 		frame:RegisterEvent('QUEST_LOG_UPDATE');
-		frame:RegisterEvent('TRADE_SKILL_CLOSE');
+		frame:RegisterEvent('TRADE_SKILL_SHOW');
 		frame:RegisterEvent('SPELLS_CHANGED');
 		frame:RegisterEvent('CHARACTER_POINTS_CHANGED');
+		frame:RegisterEvent('SKILL_LINES_CHANGED');
+		frame:RegisterEvent('PLAYER_XP_UPDATE');
+		frame:RegisterEvent('PLAYER_UPDATE_RESTING');
+		frame:RegisterEvent('PLAYER_MONEY');
 
 		frame:SetScript('OnEvent', function (self, event, bagId)
 			if (event == 'BAG_UPDATE') then
@@ -46,8 +50,8 @@ frame:SetScript('OnEvent', function (self, event, ...)
 			if (event == 'QUEST_LOG_UPDATE') then
 				updateQuests();
 			end
-			if (event == 'TRADE_SKILL_CLOSE') then
-				updateSkills();
+			if (event == 'TRADE_SKILL_SHOW') then
+				updateTradeskills();
 			end
 			if (event == 'SPELLS_CHANGED') then
 				updateSpells();
@@ -55,9 +59,22 @@ frame:SetScript('OnEvent', function (self, event, ...)
 			if (event == 'CHARACTER_POINTS_CHANGED') then
 				updateTalents();
 			end
+			if (event == 'SKILL_LINES_CHANGED') then
+				updateSkills();
+			end
+			if (event == 'PLAYER_XP_UPDATE') then
+				updateCharacter();
+			end
+			if (event == 'PLAYER_UPDATE_RESTING') then
+				updateCharacter();
+			end
+			if (event == 'PLAYER_MONEY') then
+				updateCharacter();
+			end
 		end);
 
-		initPulse();
 		updateTalents();
+		updateSkills();
+		updateCharacter();
 	end
 end);